Reviewer comments
Move free list iterators to alloc
